Resolves the rest of ticket #185
    +Letter generation now respects the assignment_notification flag
    +Fixed a bug in the sql update

+++ boost/0_3_3.sql
@@ -1,1 +1,1 @@
-alter table hms_residence_hall add column assignment_notifications integer not null default 1;
+alter table hms_residence_hall add column assignment_notifications smallint not null default 1;
